Likelihood Ratio Tests - Mixture Model - Beta-Binomial Distribution

This function can be used to estimate the most likely ploidal level using the bet

Usage

deltaquackBB(
  xm,
  trunc = c(0, 0),
  cores = NA,
  type = "alpha",
  free = "BUM",
  return = "both",
  error = NULL,
  tau = NULL
)

Arguments

xm

Matrix with two columns with total coverage and coverage for a randomly sampled allele.

trunc

List of two values representing the lower and upper bounds for allele frequency truncation , $c_L$ and $c_U$. If no truncation has been done, this should be set to c(0,0), which is the default.

cores

Number of cores to use for parallel processing.

type

Sets the free parameter for the fixed models. In nQuire, the variance is free in the fixed models. Here we allow the free parameter in the fixed model to be: 'variance', 'alpha' (the proportions of each mixture), or 'both' (variance & alpha). Options: 'variance', 'alpha', or 'both'.

free

Indicates what free model to utilize. Defaults to 'BUM', the beta-uniform model. Options: 'BUM' (Beta-uniform mixture), 'BM' (Beta mixture).

return

Indicates what values to return. Options: 'deltaLL' (delta log-likelihood) or 'BIC' (bayesian information criterion) or 'both'.

error

Sequencing error rate.

tau

Sequence overdispersion parameter for read counts.

Value

Delta log-likelihood or BIC scores for fixed models including diploid, triploid, tetraploid, pentaploid, and hexaploid. For both, the smallest score is the most likely model.